requirements.txt
PyMySQL==1.0.2
python-dotenv==1.0.1
requests==2.31.0
yoyo-migrations==8.2.0
cryptography